requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-generativeai>=0.3.0
urllib3>=2.0.0
